+++ extension/content_dashboard.js
@@ -172,10 +172,8 @@ function displayRouteManagement() {
 
                 //Open new tabs
                 for (let i = 0; i < pages.length; i++) {
+                    if (i >= 5) break;
                     window.open(pages[i], '_blank');
-                    if (i == 9) {
-                        break;
-                    }
                 }
             });