+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.0c] - 2025-05-10
+
+### Changed
+- An interval for page opening has been added for extracting flight data. [#3](https://github.com/NEWLY2014/AirlineSim-Enhancement-Suite/issues/3)
+
 ## [0.7.0b] - 2025-05-05
 
 ### Changed